- Refactored import/export of components - Refactored removeProduct function return type - Refactored reducer for remove product error
