+++ src/Components/index.ts
@@ -1,0 +1,10 @@
+import { HomePageHeader } from "./HomePageHeader/HomePageHeader";
+import { Link, Login } from "./Link/Link";
+import { ProductDetails } from "./ProductDetails/ProductDetails";
+
+export { 
+    HomePageHeader,
+    Link,
+    Login,
+    ProductDetails,
+};
