+++ src/Pages/Home/homePage.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import {Container, BigButton} from './styled';
-import {HomePageHeader} from "../../Components/HomePageHeader/HomePageHeader";
-import {Paper} from "../../Layouts/Paper/Paper";
-import {Login} from "../../Components/Link/Link";
+import { Container, BigButton } from './styled';
+import { HomePageHeader } from "../../Components";
+import { Paper } from "../../Layouts/Paper/Paper";
+import { Login } from "../../Components";
 
 let title = 'Manage your finances';
 let text = `Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore
